requirements.txt
pystray
pyaudio
Pillow
encodec
